+++ src/Resources/Resources/ChangeLog.md
@@ -21,6 +21,7 @@
 ## Upcoming Release
 * Fixed the issue that Get-AzReource not working with `-ExpandProperties`. [#11248]
 * Updated Resources SDK to 2024-11-01.
+* Added ValidationLevel Parameter to WhatIf and Validate cmdlets for deployments.
 
 ## Version 7.9.0
 * Added `-ApplicationId` as an alias of `-ServicePrincipalName`.